split replay tests into a package
